pyproject.toml
[tool.poetry]
name = "hermes"
version = "0.2"
description = "Interface de aquisição de dados geológicos, escrita em Python (3.10), utilizando a framework Django (+ DRF ) juntamente ao Prometehus e Grafana."
authors = ["Otávio Canuto"]
license = "GNU AGPLv3 "
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
prometheus-client = "^0.16.0"
prometheus-api-client = "^0.5.3"
openpyxl = "^3.1.2"
celery = "^5.2.7"
sqlalchemy = "^2.0.12"
paho-mqtt = "^1.6.1"
psutil = "^5.9.5"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"
psycopg2-binary = "^2.9.6"
watchdog = "^3.0.0"
django-compressor = "^4.3.1"
django-plotly-dash = "^2.2.1"
grafanalib = "^0.7.0"
redis = "^4.6.0"
django-widget-tweaks = "^1.4.12"
mysqlclient = "^2.2.0"
django-cacheops = "^7.0.1"
django-psdb-engine = "^1.0.6"
promql-parser = "^0.1.3"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
jupyter = "^1.0.0"
xlsxwriter = "^3.1.2"


[tool.poetry.group.node.dependencies]
python = "^3.10"
watchdog = "^3.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
